requirements.txt
openai
python-dotenv
pydantic
numpy
gradio == 5.49.1